[SPARK-43063][SQL][FOLLOWUP] Add ToPrettyString expression for Dataset.show
### What changes were proposed in this pull request?

This is a followup of https://github.com/apache/spark/pull/40699 to avoid changing the Cast behavior. It pulls out the cast-to-string code into a base trait, and add a new Expression `ToPrettyString` to extend this trait with a little customization.

It also handles binary value inside array/struct/map to also print hex format, for `df.show` only, not `Cast`.

### Why are the changes needed?

avoid behavior change

### Does this PR introduce _any_ user-facing change?

change back the behavior of casting array/map/struct to string regarding null elements. It was `null`, then changed to `NULL` in #40699 , and is `null` again after this PR.

### How was this patch tested?

existing tests

Closes #40922 from cloud-fan/pretty.

Authored-by: Wenchen Fan <wenchen@databricks.com>
